requirements.txt
pytest
python-dotenv
requests
openai
